% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_rclone_local_path.R
\name{fmt_rclone_local_path}
\alias{fmt_rclone_local_path}
\title{Format the rclone local path string}
\usage{
fmt_rclone_local_path(local_path)
}
\arguments{
\item{local_path}{(character) The rclone local path. (Default: none)}
}
\value{
(character) The rclone local path will be returned as a character
    string.
}
\description{
Return a character string of the rclone local folder path formatted for
    use as an argument to rclone, as executed with, e.g., `base::system2()`.
}
\section{Details}{

The rclone local path string is intended to be used with rclone sync.
}

\examples{
rclone_local_path <- fmt_rclone_local_path('/path/to/og_mygroup/Documents')
}
\keyword{rclone}
