% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_rclone_remote_path.R
\name{fmt_rclone_remote_path}
\alias{fmt_rclone_remote_path}
\title{Format the rclone remote path string}
\usage{
fmt_rclone_remote_path(remote_name, remote_path)
}
\arguments{
\item{remote_name}{(character) The rclone remote name. (Default: none)}

\item{remote_path}{(character) The rclone remote path. (Default: none)}
}
\value{
(character) The rclone remote path will be returned as a character
    string.
}
\description{
Return a character string of the rclone remote folder path formatted for
    use as an argument to rclone, as executed with, e.g., `base::system2()`.
}
\section{Details}{

The rclone remote path string is intended to be used with rclone sync.
}

\examples{
rclone_remote_path <- fmt_rclone_remote_path('og_mygroup', 'Documents')
}
\keyword{rclone}
