% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_m365_folders.R
\name{get_m365_folders}
\alias{get_m365_folders}
\title{Construct Microsoft 365 (m365) folder paths}
\usage{
get_m365_folders(
  m365_org = "ORG",
  m365_names = c("onedrive", "sharepoint"),
  m365_prefix = c("OneDrive - ", ""),
  parent_folder = fs::path_home()
)
}
\arguments{
\item{m365_org}{(character) The m365 organization name. (Default: 'ORG')}

\item{m365_names}{(character) Names to be assigned to the list of folders.
(Default: c('onedrive', 'sharepoint'))}

\item{m365_prefix}{(character) Folder name prefixes to used with m365_org.
(Default: c('OneDrive - ', ''))}

\item{parent_folder}{(character) Folder name under which the m365 folders
will be found. (Default: fs::path_home())}
}
\value{
(list) The named folder paths for a m365 file structure, will be
    returned as a list.
}
\description{
Return a named list of m365 folder paths that can be used for local storage.
}
\section{Details}{

The list of folders can be used to create any which are missing or to
refer to a folder path by name to avoid hardcoding paths in scripts.

The folder paths will be subfolders relative to the parent folder (usually
your home folder) and will be used for syncing a local copy with m365 files.
}

\examples{
m365_folders <- get_m365_folders(m365_org = 'UW')
}
\keyword{m365}
