% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rclone_sync.R
\name{rclone_sync}
\alias{rclone_sync}
\title{Sync with remote using rclone}
\usage{
rclone_sync(
  remote_name,
  src,
  dest,
  cmd = "rclone",
  cmd_args = c("sync", "--update", "--auto-confirm")
)
}
\arguments{
\item{remote_name}{(character) The rclone remote name. (Default: none)}

\item{src}{(character) The rclone path for the sync source. (Default: none)}

\item{dest}{(character) The rclone path for the sync destination.
(Default: none)}

\item{cmd}{(character) The rclone executable name or the path to the rclone
executable. (Default: 'rclone')}

\item{cmd_args}{(character vector) The rclone command arguments.
(Default: c('sync', '--update', '--auto-confirm'))}
}
\value{
(none or error message) An error message will be returned upon
    error, otherwise nothing is returned.
}
\description{
Sync from the source (src) to the destination (dest) using rclone sync.
}
\section{Details}{

Rclone will be called with `system2()` three times: (1) to get the path
    to the config file, (2) to search the config file for the remote name,
    and (3) to sync the source to the destination. The source and
    destination character strings can be either local or remote, and should
    be constructed with `fmt_rclone_local_path()` and
    `fmt_rclone_remote_path()`, respectively .
}

\examples{
rclone_remote_path <- fmt_rclone_remote_path('og_mygroup', 'Documents')
rclone_local_path <- fmt_rclone_local_path('/path/to/og_mygroup/Documents')
\dontrun{
rclone_sync('og_mygroup', rclone_remote_path, rclone_local_path)
}
}
\keyword{rclone}
